only allow admins to archive collections

+++ app/Http/Controllers/CollectionController.php
@@ -4,6 +4,7 @@ namespace App\Http\Controllers;
 
 use App\Http\Requests\CollectionRequest;
 use App\Models\Collection;
+use Illuminate\Http\RedirectResponse;
 use Illuminate\Http\Request;
 
 class CollectionController extends Controller
@@ -99,11 +100,22 @@ class CollectionController extends Controller
     /**
      * Remove the specified resource from storage.
      *
-     * @param  \App\Models\Collection  $collection
-     * @return \Illuminate\Http\Response
+     * @param Request $request
+     * @param Collection $collection
+     * @return RedirectResponse
      */
-    public function destroy(Collection $collection)
+    public function destroy(Request $request, Collection $collection): RedirectResponse
     {
+        if ($request->user()->cannot('delete', $collection)) {
+            return redirect()->back()
+                ->with('notification', [
+                    'actions' => false,
+                    'message' => 'You do not have the required permissions to archive this collection.',
+                    'title' => 'Access error',
+                    'type' => 'error',
+                ]);
+        }
+
         $collection->delete();
 
         return redirect()->route('editor.collections.index');
